move query definitions
